Huge refactor to make processing device messages easier. Added Encounter model.

+++ app/models/device_message_processor.rb
@@ -1,6 +1,5 @@
 # Given a device message, creates the associated messages in the DB with their samples, and indexes them
 class DeviceMessageProcessor
-
   attr_reader :device_message
 
   def initialize device_message
@@ -25,9 +24,7 @@ class DeviceMessageProcessor
     @device_message.institution
   end
 
-
   class SingleMessageProcessor
-
     attr_reader :parsed_message, :parent
 
     delegate :device, :device_message, :client, to: :parent
@@ -41,24 +38,26 @@ class DeviceMessageProcessor
       test = find_or_initialize_test
       is_new_test = test.new_record?
       old_sample = test.sample
-      old_patient = test.current_patient
+      old_patient = test.patient
+      old_encounter = test.encounter
 
-      process_sample test
-      process_patient test
+      encounter_indexed_fields = process_encounter test
+      sample_indexed_fields    = process_sample test
+      patient_indexed_fields   =  process_patient test
 
-      sample_changed = test.sample_id_changed?
-      patient_changed = test.patient_id_changed? ||
-        (test.sample.present? && test.sample.patient_id_changed?)
+      save_encounter test, encounter_indexed_fields
+      save_sample    test, sample_indexed_fields
+      save_patient   test, patient_indexed_fields
+
+      encounter_changed = test.encounter_id_changed?
+      sample_changed    = test.sample_id_changed?
+      patient_changed   = test.patient_id_changed?
 
       test.save!
 
-      if sample_changed && old_sample.present?
-        old_sample.destroy
-      end
-
-      if patient_changed && old_patient.present?
-        old_patient.destroy
-      end
+      old_encounter.destroy  if encounter_changed && old_encounter
+      old_sample.destroy     if sample_changed    && old_sample
+      old_patient.destroy    if patient_changed   && old_patient
 
       index_test test, is_new_test
     end
@@ -66,18 +65,13 @@ class DeviceMessageProcessor
     private
 
     def find_or_initialize_test
-      pii = {"test" => parsed_message["test"]["pii"]}
-      custom_fields = {"test" => parsed_message["test"]["custom"]}
-      indexed_fields = {"test" => parsed_message["test"]["indexed"]}
-
+      test_id = parsed_message["test"]["indexed"]["id"]
       test = TestResult.new device_messages: [device_message],
-                            plain_sensitive_data: pii, # TODO: this should be what the manifest returns
-                            custom_fields: custom_fields,
-                            indexed_fields: indexed_fields,
-                            test_id: indexed_fields["test"]["id"],
+                            test_id: test_id,
                             device: device
+      assign_fields parsed_message, test, "test"
 
-      if test.test_id && existing = TestResult.find_by(test_id: test.test_id, device_id: test.device_id)
+      if test_id && existing = TestResult.find_by(test_id: test_id, device_id: test.device_id)
         existing.merge(test)
         existing
       else
@@ -86,103 +80,108 @@ class DeviceMessageProcessor
     end
 
     def find_or_initialize_sample
-      pii = {"sample" => parsed_message["sample"]["pii"]}
-      custom_fields = {"sample" => parsed_message["sample"]["custom"]}
-      indexed_fields = {"sample" => parsed_message["sample"]["indexed"]}
-      sample_uid = pii["sample"]["uid"]
-
-      sample = Sample.new plain_sensitive_data: pii,
-                          custom_fields: custom_fields,
-                          indexed_fields: indexed_fields,
-                          institution_id: @parent.institution.id
-
-      if sample_uid.present? && existing = Sample.find_by_pii(sample_uid, @parent.institution.id)
-        existing_indexed = existing.indexed_fields.deep_dup
-        existing.merge(sample)
-        [existing, existing_indexed]
-      else
-        [sample, nil]
-      end
+      find_or_initialize_entity Sample, "sample", ((parsed_message["sample"] || {})["pii"] || {})["uid"]
     end
 
     def find_or_initialize_patient
-      pii = {"patient" => parsed_message["patient"]["pii"]}
-      custom_fields = {"patient" => parsed_message["patient"]["custom"]}
-      indexed_fields = {"patient" => parsed_message["patient"]["indexed"]}
-      patient_id = pii["patient"]["id"]
+      find_or_initialize_entity Patient, "patient", ((parsed_message["patient"] || {})["pii"] || {})["id"]
+    end
 
-      patient = Patient.new plain_sensitive_data: pii,
-                            custom_fields: custom_fields,
-                            indexed_fields: indexed_fields,
-                            institution_id: @parent.institution.id
+    def find_or_initialize_encounter
+      find_or_initialize_entity Encounter, "encounter", ((parsed_message["encounter"] || {})["pii"] || {})["id"]
+    end
 
-      if patient_id.present? && existing = Patient.find_by_pii(patient_id, @parent.institution.id)
+    def find_or_initialize_entity(klass, scope, entity_id)
+      new_entity = klass.new institution_id: @parent.institution.id
+      assign_fields parsed_message, new_entity, scope
+
+      if entity_id && (existing = klass.find_by_pii(entity_id, @parent.institution.id))
         existing_indexed = existing.indexed_fields.deep_dup
-        existing.merge(patient)
+        existing.merge(new_entity)
         [existing, existing_indexed]
       else
-        [patient, nil]
+        [new_entity, nil]
       end
+    end
+
+    def process_encounter(test)
+      encounter, existing_indexed = find_or_initialize_encounter
+
+      if test.encounter
+        if encounter.encounter_id && (encounter.encounter_id != test.encounter.encounter_id)
+          test.encounter = encounter
+        else
+          test.encounter.merge_entity_scope encounter, "encounter"
+        end
+      else
+        if encounter.encounter_id
+          test.encounter = encounter
+          test.move_entity_scope "encounter", test.encounter
+        else
+          test.merge_entity_scope encounter, "encounter"
+        end
+      end
+
+      existing_indexed
     end
 
     def process_sample(test)
       sample, existing_indexed = find_or_initialize_sample
 
-      if test.sample.present?
-        if sample.sample_uid.present?
-          if test.sample.sample_uid == sample.sample_uid
-            test.sample.merge sample
-          else
-            test.sample = sample
-          end
+      if test.sample
+        if sample.sample_uid && (sample.sample_uid != test.sample.sample_uid)
+          test.sample = sample
         else
-          test.sample.merge sample
+          test.sample.merge_entity_scope sample, "sample"
         end
       else
-        if sample.sample_uid.present?
-          test.extract_sample_data_into sample
+        if sample.sample_uid
           test.sample = sample
-          test.patient = sample.patient unless test.patient.present?
+          test.move_entity_scope "sample", test.sample
         else
-          test.add_sample_data sample
+          test.merge_entity_scope sample, "sample"
         end
       end
 
-      test.sample.save! if test.sample.present?
-
-      if !existing_indexed.nil? && existing_indexed != test.sample.try(:indexed_fields)
-        update_sample_in_existing_documents_with test.sample
-      end
+      existing_indexed
     end
 
     def process_patient(test)
       patient, existing_indexed = find_or_initialize_patient
-      current = test.current_patient
 
-      if current.present?
-        if patient.patient_id.present?
-          if current.patient_id == patient.patient_id
-            current.merge patient
-          else
-            test.current_patient = patient
-          end
+      if test.patient
+        if patient.patient_id && (patient.patient_id != test.patient.patient_id)
+          test.patient = patient
         else
-          current.merge patient
+          test.patient.merge_entity_scope patient, "patient"
         end
       else
-        if patient.patient_id.present?
-          test.extract_patient_data_into patient
-          test.current_patient = patient
+        if patient.patient_id
+          test.patient = patient
+          test.move_entity_scope "patient", test.patient
         else
-          test.add_patient_data patient
+          test.merge_entity_scope patient, "patient"
         end
       end
 
-      test.current_patient.save! if test.current_patient.present?
+      # Move patient data to sample if there's one,
+      # or move sample's patient to test, if the test has no patient
+      if test.sample
+        test.move_entity_scope "patient", test.sample
 
-      if !existing_indexed.nil? && existing_indexed != test.current_patient.try(:indexed_fields)
-        update_patient_in_existing_documents_with test.current_patient
+        if test.patient
+          test.sample.move_entity_scope "patient", test.patient
+          test.sample.patient = test.patient
+        else
+          test.sample.merge_entity_scope patient, "patient"
+          if test.sample.patient
+            test.patient = test.sample.patient
+            test.sample.patient.merge_entity_scope patient, "patient"
+          end
+        end
       end
+
+      existing_indexed
     end
 
     def index_test(test, is_new)
@@ -190,24 +189,43 @@ class DeviceMessageProcessor
       is_new ? indexer.index : indexer.update
     end
 
-    def update_sample_in_existing_documents_with sample
-      response = client.search index: Cdx::Api.index_name, body:{query: { filtered: { filter: { term: { :'sample.uuid' => sample.uuid } } } }, fields: []}, size: 10000
+    def assign_fields(parsed_message, entity, scope)
+      entity.indexed_fields       = {scope => (parsed_message[scope] || {})["indexed"]}
+      entity.custom_fields        = {scope => (parsed_message[scope] || {})["custom"]}
+      entity.plain_sensitive_data = {scope => (parsed_message[scope] || {})["pii"]}
+    end
+
+    def update_entity_in_elasticsearch(entity, key)
+      response = client.search index: Cdx::Api.index_name, body:{query: { filtered: { filter: { term: { key => entity.uuid } } } }, fields: []}, size: 10000
       body = response["hits"]["hits"].map do |element|
-        { update: { _type: element["_type"], _id: element["_id"], data: { doc: sample.indexed_fields } } }
+        { update: { _type: element["_type"], _id: element["_id"], data: { doc: entity.indexed_fields } } }
       end
 
       client.bulk index: Cdx::Api.index_name, body: body unless body.blank?
     end
 
-    def update_patient_in_existing_documents_with patient
-      response = client.search index: Cdx::Api.index_name, body:{query: { filtered: { filter: { term: { :'patient.uuid' => patient.uuid } } } }, fields: []}, size: 10000
-      body = response["hits"]["hits"].map do |element|
-        { update: { _type: element["_type"], _id: element["_id"], data: { doc: patient.indexed_fields } } }
-      end
+    def save_encounter(test, old_indexed_fields)
+      test.encounter.try(:save!)
 
-      client.bulk index: Cdx::Api.index_name, body: body unless body.blank?
+      if old_indexed_fields && old_indexed_fields != test.encounter.try(:indexed_fields)
+        update_entity_in_elasticsearch test.encounter, 'encounter.uuid'
+      end
     end
 
+    def save_sample(test, old_indexed_fields)
+      test.sample.try(:save!)
+
+      if old_indexed_fields && old_indexed_fields != test.sample.try(:indexed_fields)
+        update_entity_in_elasticsearch test.sample, 'sample.uuid'
+      end
+    end
+
+    def save_patient(test, old_indexed_fields)
+      test.patient.try(:save!)
+
+      if old_indexed_fields && old_indexed_fields != test.patient.try(:indexed_fields)
+        update_entity_in_elasticsearch test.patient, 'patient.uuid'
+      end
+    end
   end
-
 end
